+++ src/main/java/fr/maxlego08/essentials/commands/commands/teleport/CommandTeleport.java
@@ -37,11 +37,10 @@ public class CommandTeleport extends VCommand {
             Location location = player.getLocation();
             float yaw, pitch;
 
-            if (isNumeric(value)) {
-
-                double x = this.argAsDouble(0);
-                double y = this.argAsDouble(1);
-                double z = this.argAsDouble(2);
+            if (isNumeric(value) || isRelativeCoordinate(value)) {
+                double x = parseCoordinate(value, location.getX());
+                double y = parseCoordinate(this.argAsString(1), location.getY());
+                double z = parseCoordinate(this.argAsString(2), location.getZ());
                 yaw = (float) this.argAsDouble(3, location.getYaw());
                 pitch = (float) this.argAsDouble(4, location.getPitch());
 
@@ -53,9 +52,9 @@ public class CommandTeleport extends VCommand {
                 message(this.sender, Message.COMMAND_TP_LOCATION, "%x%", x, "%y%", y, "%z%", z);
             } else {
 
-                double x = this.argAsDouble(1);
-                double y = this.argAsDouble(2);
-                double z = this.argAsDouble(3);
+                double x = parseCoordinate(this.argAsString(1), location.getX());
+                double y = parseCoordinate(this.argAsString(2), location.getY());
+                double z = parseCoordinate(this.argAsString(3), location.getZ());
 
                 location.set(x, y, z);
 
@@ -96,5 +95,21 @@ public class CommandTeleport extends VCommand {
             return false;
         }
         return true;
+    }
+
+    private boolean isRelativeCoordinate(String coordinate) {
+        return coordinate != null && coordinate.startsWith("~");
+    }
+
+    private double parseCoordinate(String coordinate, double current) {
+        if (coordinate == null) return current;
+        if (coordinate.startsWith("~")) {
+            if (coordinate.length() == 1) {
+                return current;
+            } else {
+                return current + Double.parseDouble(coordinate.substring(1));
+            }
+        }
+        return Double.parseDouble(coordinate);
     }
 }