+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Added ``/clearinventory [<player>]`` [#101](https://github.com/Maxlego08/zEssentials/issues/124)
 - Improve economy module with offline players
+- Fixed teleport command with relative coordinates [#142](https://github.com/Maxlego08/zEssentials/issues/142)
 
 # 1.0.1.4